refactor: remove SuggestionsDbio
